Added color styles; a bit more branding.
Added buttons to the homepage.

Updated some images.

Added some placeholder images.

Added a contact us area to the bottom of the page with working links to call and get directions.

Made sure that the when navigating the site that the user is brought back to the top of the page.

Corrected information in the JSON-LD.